button debounce circuit fixed and solder jumpers added
Solder jumpers added to self destruct button's vcu connection. This is because when the button is used for changing lcd information, it has nothing to do with the VCU so needs no connection. Also prevents the VCU accidentally damaging the dash board through these pins.

Also, minor silkscreen placement improvements.